+++ config/production.yaml
@@ -5,3 +5,4 @@ log_level: INFO
 service:
   host: 0.0.0.0
   port: 6333
+  grpc_port: 6334
