+++ src/actix/api/collections_api.rs
@@ -1,13 +1,14 @@
-use crate::api::models::{CollectionDescription, CollectionsResponse};
-use crate::common::helpers::process_response;
+use crate::actix::api::models::{CollectionDescription, CollectionsResponse};
+use crate::actix::helpers::process_response;
 use actix_web::rt::time::Instant;
 use actix_web::{get, post, web, Responder};
 use itertools::Itertools;
+use std::sync::Arc;
 use storage::content_manager::storage_ops::StorageOperations;
 use storage::content_manager::toc::TableOfContent;
 
 #[get("/collections")]
-pub async fn get_collections(toc: web::Data<TableOfContent>) -> impl Responder {
+pub async fn get_collections(toc: web::Data<Arc<TableOfContent>>) -> impl Responder {
     let timing = Instant::now();
 
     let response = {
@@ -25,7 +26,7 @@ pub async fn get_collections(toc: web::Data<TableOfContent>) -> impl Responder {
 
 #[get("/collections/{name}")]
 pub async fn get_collection(
-    toc: web::Data<TableOfContent>,
+    toc: web::Data<Arc<TableOfContent>>,
     path: web::Path<String>,
 ) -> impl Responder {
     let name = path.into_inner();
@@ -41,7 +42,7 @@ pub async fn get_collection(
 
 #[post("/collections")]
 pub async fn update_collections(
-    toc: web::Data<TableOfContent>,
+    toc: web::Data<Arc<TableOfContent>>,
     operation: web::Json<StorageOperations>,
 ) -> impl Responder {
     let timing = Instant::now();
